ML_Workshop_Data/XSUAASample/XsuaaDemoService/requirements.txt
flask
flask_cors
flask-jwt
cryptography
pycrypto
pyjWT




